+++ src/main/org/firebirdsql/gds/ng/FbStatement.java
@@ -119,7 +119,8 @@ public interface FbStatement extends ExceptionListenable {
      * implementations may opt to not close the cursor on the server if the server closes the cursor automatically.
      * </p>
      *
-     * @param transactionEnd Close is in response to a transaction end.
+     * @param transactionEnd
+     *         Close is in response to a transaction end.
      * @throws SQLException
      */
     void closeCursor(boolean transactionEnd) throws SQLException;
@@ -282,10 +283,46 @@ public interface FbStatement extends ExceptionListenable {
      * cursor.
      * </p>
      *
-     * @param transactionEnd Close is in response to a transaction end
-     *
-     * @throws SQLException If this statement is closed or the cursor could not be closed.
+     * @param transactionEnd
+     *         Close is in response to a transaction end
+     * @throws SQLException
+     *         If this statement is closed or the cursor could not be closed.
      * @since 3.0.6
      */
     void ensureClosedCursor(boolean transactionEnd) throws SQLException;
+
+    /**
+     * Sets the statement timeout.
+     * <p>
+     * The statement timeout value is ignored in implementations that do not support timeouts. If the provided
+     * timeout value is greater than supported (eg greater than ‭4294967295‬ milliseconds on Firebird 4), the
+     * implementation should behave as if zero ({@code 0}) was set, but still report the original value.
+     * </p>
+     * <p>
+     * The configured timeout only affects subsequent executes on this statement. The timeout includes time spent
+     * reading the result set.
+     * </p>
+     *
+     * @param timeoutMillis
+     *         Timeout value in milliseconds
+     * @throws SQLException
+     *         If the value is less than zero, this statement is closed, or a database access error occurs
+     * @since 4.0
+     */
+    void setTimeout(long timeoutMillis) throws SQLException;
+
+    /**
+     * Gets the current statement timeout for this statement.
+     * <p>
+     * This method will only return the current statement timeout value for this method, it will not consider attachment
+     * or connection level timeouts. This is an implementation decision that might change in a point release.
+     * </p>
+     *
+     * @return The configured timeout in milliseconds; read the documentation in {@link #setTimeout(long)}
+     * @throws SQLException
+     *         If this statement is closed, or a database access error occurs
+     * @see #setTimeout(long)
+     * @since 4.0
+     */
+    long getTimeout() throws SQLException;
 }