+++ src/main/org/firebirdsql/gds/ng/wire/version16/V16Database.java
@@ -16,30 +16,30 @@
  *
  * All rights reserved.
  */
-package org.firebirdsql.gds.ng.wire.version13;
+package org.firebirdsql.gds.ng.wire.version16;
 
-import org.firebirdsql.common.rules.RequireProtocol;
-import org.firebirdsql.gds.ng.wire.version12.TestV12Statement;
-import org.junit.ClassRule;
-
-import static org.firebirdsql.common.rules.RequireProtocol.requireProtocolVersion;
+import org.firebirdsql.gds.ng.wire.ProtocolDescriptor;
+import org.firebirdsql.gds.ng.wire.WireDatabaseConnection;
+import org.firebirdsql.gds.ng.wire.version15.V15Database;
 
 /**
- * Tests for {@link org.firebirdsql.gds.ng.wire.version13.V13Statement} in the V13 protocol, reuses test for V12.
+ * {@link org.firebirdsql.gds.ng.wire.FbWireDatabase} implementation for the version 16 wire protocol.
  *
  * @author <a href="mailto:mrotteveel@users.sourceforge.net">Mark Rotteveel</a>
- * @since 3.0
+ * @since 4.0
  */
-public class TestV13Statement extends TestV12Statement {
+public class V16Database extends V15Database {
 
-    @ClassRule
-    public static final RequireProtocol requireProtocol = requireProtocolVersion(13);
-
-    public TestV13Statement() {
-        this(new V13CommonConnectionInfo());
-    }
-
-    protected TestV13Statement(V13CommonConnectionInfo commonConnectionInfo) {
-        super(commonConnectionInfo);
+    /**
+     * Creates a V15Database instance.
+     *
+     * @param connection
+     *         A WireConnection with an established connection to the server.
+     * @param descriptor
+     *         The ProtocolDescriptor that created this connection (this is
+     */
+    protected V16Database(WireDatabaseConnection connection,
+            ProtocolDescriptor descriptor) {
+        super(connection, descriptor);
     }
 }