+++ src/test/org/firebirdsql/gds/ng/wire/version15/TestV15Statement.java
@@ -25,7 +25,7 @@ import org.junit.ClassRule;
 import static org.firebirdsql.common.rules.RequireProtocol.requireProtocolVersion;
 
 /**
- * Tests for {@link org.firebirdsql.gds.ng.wire.version12.V12Statement} in the V15 protocol, reuses test for V13.
+ * Tests for {@link org.firebirdsql.gds.ng.wire.version13.V13Statement} in the V15 protocol, reuses test for V13.
  *
  * @author <a href="mailto:mrotteveel@users.sourceforge.net">Mark Rotteveel</a>
  * @since 4.0